requirements.txt

crewai==0.30.11
crewai-tools==0.2.6
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==1.0.3
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
google-generativeai==0.5.4
google-genai==1.16.0
openai==1.79.0


supabase==2.15.2
postgrest==1.0.2
storage3==0.11.3
gotrue==2.12.0
realtime==2.4.3
supafunc==0.9.4
psycopg2-binary==2.9.10
pgvector==0.3.6


PyPDF2==3.0.1
pypdf==4.3.1
PyMuPDF==1.26.0

python-dotenv==1.0.1

requests==2.32.3
httpx==0.28.1
aiohttp==3.11.18

google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.93.0
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0

pandas==2.2.3
numpy==1.26.4

tiktoken==0.7.0
tokenizers==0.21.1

tenacity==8.5.0
retry==0.9.2

tqdm==4.67.1
coloredlogs==15.0.1

pydantic==2.11.4
PyYAML==6.0.2
click==8.1.7
rich==13.9.4

pytest==8.3.5
pytest-mock==3.14.1 